requirements.txt
pyserial
pycryptodome
paho-mqtt
python-decouple